feat: Add conditional logic for updating events in EventForm
The code changes in  add conditional logic to handle updating events. If an  object is provided, the code sends a PATCH request to update the event with the new data. Otherwise, it sends a POST request to create a new event. This improves the functionality of the form by allowing users to update existing events.