feat: Add delete functionality to DeleteButton component
This commit adds the functionality to delete events using the DeleteButton component. When the \Delete\ button is clicked, a DELETE request is sent to the server to delete the event with the corresponding event ID. After successful deletion, the user is redirected to the events page. This enhancement improves the user experience by providing a convenient way to remove unwanted events from the application.