fix: Fix optional chaining in EventForm and EventDetails components
This commit fixes the optional chaining in the EventForm and EventDetails components to correctly access the location and date properties of the event object. The previous code used incorrect optional chaining syntax, which caused issues with displaying the location and date. This fix improves the accuracy and reliability of the displayed event information.